requirements.txt
streamlit
pandas
pdfminer.six
docx2txt
spacy
nltk
scikit-learn
sentence-transformers
torch